split git restore and http restore
